fetch/bugs/request-etag.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Request ETag</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#request">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    promise_test(function() {
      var cacheBuster = Math.random(); // ensures first request is uncached
      var endpoint = "resources/cache.py?v=" + cacheBuster;

      // make the first request
      return fetch(endpoint).then(function(response) {
        // ensure we're getting the regular, uncached response
        assert_equals(response.status, 200);

        return response.text(); // consuming the body, just to be safe
      }).then(function(body) {
        // make a second request
        return fetch(endpoint);
      }).then(function(response) {
        // a 304 response indicates our browser sent the correct
        // If-None-Match request header
        assert_equals(response.status, 304);

        return response.text(); // consuming the body, just to be safe
      });
    }, "Testing conditional GET with ETags");

    done();
    </script>
  </body>
</html>
